requirements.txt
Django==2.2.5
djangorestframework==3.10.3
psycopg2==2.8.3
drf-yasg==1.16.1
